refactor(problems/slug): componentize, clean up code
